+++ BambooTracker/gui/mainwindow.cpp
@@ -53,7 +53,8 @@ MainWindow::MainWindow(std::weak_ptr<Configuration> config, QString filePath, QW
 	isEditedPattern_(true),
 	isEditedOrder_(false),
 	isEditedInstList_(false),
-	isSelectedPO_(false)
+	isSelectedPO_(false),
+	effListDiag_(std::make_unique<EffectListDialog>())
 {
 	ui->setupUi(this);
 
@@ -2320,4 +2321,10 @@ void MainWindow::updateVisuals()
 	bt_->getOutputHistory(wave);
 
 	ui->waveVisual->setStereoSamples(wave, OPNAController::OutputHistorySize);
+}
+
+void MainWindow::on_action_Effect_List_triggered()
+{
+	if (effListDiag_->isVisible()) effListDiag_->activateWindow();
+	else effListDiag_->show();
 }